Turn render-summary.py into a `limbo` command
One less script floating around the repo.
