+++ .github/workflows/test-harness.yml
@@ -24,8 +24,7 @@ jobs:
         run: make test-go
 
       - name: render summary
-        run: |
-          python ./harness/render-summary.py limbo.json results.json
+        run: make run ARGS="render-sumary"
   openssl:
     runs-on: ubuntu-latest
     steps:
@@ -39,8 +38,7 @@ jobs:
         run: make CXX=clang++ test-openssl
 
       - name: render summary
-        run: |
-          python ./harness/render-summary.py limbo.json results.json
+        run: make run ARGS="render-sumary"
 
   rust-webpki:
     runs-on: ubuntu-latest
@@ -55,8 +53,7 @@ jobs:
         run: make test-rust-webpki
 
       - name: render summary
-        run: |
-          python ./harness/render-summary.py limbo.json results.json
+        run: make run ARGS="render-sumary"
 
   rustls-webpki:
     runs-on: ubuntu-latest
@@ -71,6 +68,5 @@ jobs:
         run: make test-rustls-webpki
 
       - name: render summary
-        run: |
-          python ./harness/render-summary.py limbo.json results.json
+        run: make run ARGS="render-sumary"
